+++ tests/dummy/config/json-api-v1.php
@@ -58,6 +58,7 @@ return [
         'downloads' => \DummyApp\Download::class,
         'phones' => \DummyApp\Phone::class,
         'posts' => \DummyApp\Post::class,
+        'queue-jobs' => \CloudCreativity\LaravelJsonApi\Queue\ClientJob::class,
         'sites' => \DummyApp\Entities\Site::class,
         'tags' => \DummyApp\Tag::class,
         'users' => \DummyApp\User::class,
@@ -151,7 +152,6 @@ return [
     |
     */
     'providers' => [
-        \CloudCreativity\LaravelJsonApi\Resources\ResourceProvider::class,
         \DummyPackage\ResourceProvider::class,
     ],
